LLM/Code/weather-api/requirements.txt
typer
requests
python-dotenv
